# docs/swagger.yaml
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  models.Game:
    properties:
      id:
        type: integer
      name:
        type: string
      percentage:
        minLength: 2
        type: string
      plataform:
        type: string
      status:
        minLength: 1
        type: string
    type: object
info:
  contact: {}
paths:
  /games:
    delete:
      consumes:
      - application/json
      description: With params Name, Plataform, Status, Percentage removes a new game
      parameters:
      - description: Game Model
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Removes a new game on database
      tags:
      - Games
    get:
      consumes:
      - application/json
      description: Route to show all games
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Show all games
      tags:
      - Games
    patch:
      consumes:
      - application/json
      description: With params Name Plataform, Status and Percentage updates game
        state
      parameters:
      - description: Game Model
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Update game status and parcentage
      tags:
      - Games
    post:
      consumes:
      - application/json
      description: With params Name, Plataform, Status, Percentage adds a new game
      parameters:
      - description: Game Model
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Adds a new game on database
      tags:
      - Games
  /games/:name/:plataform:
    get:
      consumes:
      - application/json
      description: Route to show a game with params name and plataform
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Show a game
      tags:
      - Games
swagger: "2.0"